feat: add validation when no allowed payment methods selected
